fixed controllers and view, fully implemented back end

+++ DMS/app/src/main/java/com/example/kbb12/dms/FitnessInfo/FitnessInfoView.java
@@ -25,6 +25,6 @@ public class FitnessInfoView implements ModelObserver {
 
     @Override
     public void update() {
-        mTVCal.setText(Double.toString(lmodel.getCalCount(context)));
+        mTVCal.setText(Integer.toString(lmodel.getCalCount(context)));
     }
 }
